vst_version intptr_t -> int
git-svn-id: https://svn.code.sf.net/p/fsthost/code/trunk@378 75a15a69-0cfc-47d9-a1bd-ce2941b697b2

+++ fst/fst.h
@@ -117,7 +117,7 @@ typedef struct _FST {
 	bool			program_changed;
 
 	/* Info */
-	intptr_t		vst_version;
+	int			vst_version;
 	bool			isSynth;
 	bool			canReceiveVstEvents;
 	bool			canReceiveVstMidiEvent;
